python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow>=10.2.0
tensorflow==2.10.1
numpy>=1.20,<1.24
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
